Update sign-up channel creation flow
The old flow have the ability to let the user indicate whether they are interested in YT Sync early on in the form, but for Oauth, we are starting with the SSO portal first, so we can't pre-decide which card to show when we reach the "create channel" stage.

Simplify it by just showing the regular channel creation card, which has a reminder below asking if they want to sync YT instead. Users can manually click it. Even if they miss it, they can still sync after the fact.